+++ hypervisor/include/arch/x86/guest/gpr.h
@@ -7,25 +7,4 @@
 #ifndef _VGPR_H_
 #define _VGPR_H_
 
-/* General-purpose register layout aligned with the general-purpose register idx
- * when vmexit, such as vmexit due to CR access, refer to SMD Vol.3C 27-6.
- */
-struct acrn_gp_regs {
-	uint64_t rax;
-	uint64_t rcx;
-	uint64_t rdx;
-	uint64_t rbx;
-	uint64_t rsp;
-	uint64_t rbp;
-	uint64_t rsi;
-	uint64_t rdi;
-	uint64_t r8;
-	uint64_t r9;
-	uint64_t r10;
-	uint64_t r11;
-	uint64_t r12;
-	uint64_t r13;
-	uint64_t r14;
-	uint64_t r15;
-};
 #endif
