+++ hypervisor/include/arch/x86/guest/vm0_boot.h
@@ -45,46 +45,6 @@
 #define BOOT_CTX_EFER_HIGH_OFFSET   204U
 #define SIZE_OF_BOOT_CTX            296U
 
-/* struct to define how the descriptor stored in memory.
- * Refer SDM Vol3 3.5.1 "Segment Descriptor Tables"
- * Figure 3-11
- */
-struct acrn_descriptor_ptr {
-	uint16_t limit;
-	uint64_t base;
-	uint16_t reserved[3];   /* align struct size to 64bit */
-} __attribute__((packed));
-
-struct acrn_vcpu_regs {
-	struct acrn_gp_regs gprs;
-	struct acrn_descriptor_ptr gdt;
-	struct acrn_descriptor_ptr idt;
-
-	uint64_t        rip;
-	uint64_t        cs_base;
-	uint64_t        cr0;
-	uint64_t        cr4;
-	uint64_t        cr3;
-	uint64_t        ia32_efer;
-	uint64_t        rflags;
-	uint64_t        reserved_64[4];
-
-	uint32_t        cs_ar;
-	uint32_t        reserved_32[4];
-
-	/* don't change the order of following sel */
-	uint16_t        cs_sel;
-	uint16_t        ss_sel;
-	uint16_t        ds_sel;
-	uint16_t        es_sel;
-	uint16_t        fs_sel;
-	uint16_t        gs_sel;
-	uint16_t        ldt_sel;
-	uint16_t        tr_sel;
-
-	uint16_t        reserved_16[4];
-};
-
 #ifdef CONFIG_EFI_STUB
 struct efi_context {
 	struct acrn_vcpu_regs vcpu_regs;
